* Rename AzureLoginService to AzureOAuth2Service. * Throw exception if there is no user id matched with azure mailNickname.
