+++ src/Cms/Auth/LoginService.php
@@ -2,6 +2,7 @@
 
 namespace Ridibooks\Platform\Cms\Auth;
 
+use Ridibooks\Platform\Cms\Lib\AzureOAuth2Service;
 use Ridibooks\Library\CouchbaseSessionHandler;
 
 class LoginService
@@ -33,19 +34,24 @@ class LoginService
 
 	public function doAzureLoginAction($code, $azure_config)
 	{
-		$tokenOutput = AzureLoginService::requestAccessToken($code, $azure_config);
+		$tokenOutput = AzureOAuth2Service::requestAccessToken($code, $azure_config);
 		$token_type = $tokenOutput->token_type;
 		$access_token = $tokenOutput->access_token;
 		if (!$token_type || !$access_token) {
 			throw new \Exception("[requestAccessToken]\n $tokenOutput->error: $tokenOutput->error_description");
 		}
 
-		$resourceOutput = AzureLoginService::requestResource($token_type, $access_token, $azure_config);
+		$resourceOutput = AzureOAuth2Service::requestResource($token_type, $access_token, $azure_config);
 		if ($error = $resourceOutput->{'odata.error'}) {
 			throw new \Exception("[requestResource]\n $error->code: {$error->message->value}");
 		}
 
 		$id = $resourceOutput->mailNickname;
+		$user = AdminUserService::getUser($id);
+		if (!$user || $user['is_use'] != '1') {
+			throw new \Exception('잘못된 계정정보입니다.');
+		}
+
 		$this->setSessions($id);
 	}
